+++ go/sciond/server_test.go
@@ -32,6 +32,7 @@ import (
 	"github.com/scionproto/scion/go/lib/crypto/trc"
 	"github.com/scionproto/scion/go/lib/infra/modules/trust/trustdb"
 	"github.com/scionproto/scion/go/lib/log"
+	"github.com/scionproto/scion/go/lib/overlay"
 	"github.com/scionproto/scion/go/lib/sciond"
 	"github.com/scionproto/scion/go/lib/topology"
 	"github.com/scionproto/scion/go/lib/xtest"
@@ -158,7 +159,7 @@ func TestIFInfo(t *testing.T) {
 				RawEntries: []sciond.IFInfoReplyEntry{
 					{
 						IfID:     ifids[0],
-						HostInfo: servers.MakeBRHostInfos(topo.Overlay, topo.BR, topo.IFInfoMap)[0],
+						HostInfo: MakeBRHostInfos(topo.Overlay, topo.BR, topo.IFInfoMap)[0],
 					},
 				},
 			},
@@ -170,7 +171,7 @@ func TestIFInfo(t *testing.T) {
 				RawEntries: []sciond.IFInfoReplyEntry{
 					{
 						IfID:     ifids[0],
-						HostInfo: servers.MakeBRHostInfos(topo.Overlay, topo.BR, topo.IFInfoMap)[0],
+						HostInfo: MakeBRHostInfos(topo.Overlay, topo.BR, topo.IFInfoMap)[0],
 					},
 				},
 			},
@@ -329,4 +330,17 @@ func TestMain(m *testing.M) {
 		log.Root().SetHandler(log.DiscardHandler())
 	}
 	os.Exit(m.Run())
+}
+
+func MakeBRHostInfos(ot overlay.Type, brMap map[string]topology.BRInfo,
+	ifInfoMap map[common.IFIDType]topology.IFInfo) []sciond.HostInfo {
+
+	hostInfos := make([]sciond.HostInfo, 0, len(brMap))
+	for _, brInfo := range brMap {
+		// One IFID is enough to find the unique internal address. Panic if no
+		// IFIDs exist.
+		ifid := brInfo.IFIDs[0]
+		hostInfos = append(hostInfos, servers.TopoAddrToHostInfo(ot, *ifInfoMap[ifid].InternalAddr))
+	}
+	return hostInfos
 }
